Move out commonly used functions/types to analysis
