+++ src/pixelator/analysis/permute.py
@@ -10,7 +10,7 @@ import pandas as pd
 from numpy.random import Generator as RandomNumberGenerator
 from numpy.random import default_rng
 
-from pixelator.analysis.colocalization.types import (
+from pixelator.analysis.types import (
     RegionByCountsDataFrame,
 )
