+++ src/pixelator/analysis/types.py
@@ -1,0 +1,8 @@
+"""Module with types related to analysis.
+
+Copyright © 2024 Pixelgen Technologies AB.
+"""
+
+import pandas as pd
+
+RegionByCountsDataFrame = pd.DataFrame
