+++ tests/analysis/colocalization/test_estimate.py
@@ -12,12 +12,12 @@ from pixelator.analysis.colocalization.estimate import (
     estimate_observation_statistics,
     permutation_analysis_results,
 )
-from pixelator.analysis.colocalization.permute import permutations
 from pixelator.analysis.colocalization.statistics import (
     Jaccard,
     Pearson,
     apply_multiple_stats,
 )
+from pixelator.analysis.permute import permutations
 
 
 def test_estimate_observation_statistics():
