+++ tests/analysis/colocalization/test_permute.py
@@ -7,7 +7,7 @@ Copyright © 2023 Pixelgen Technologies AB.
 import pandas as pd
 from numpy.random import default_rng
 from pandas.testing import assert_series_equal
-from pixelator.analysis.colocalization.permute import (
+from pixelator.analysis.permute import (
     permutations,
     permute,
 )
